+++ frontend/src/contexts/punchlines.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useCrud from "../../hooks/useCrud";
+import useCrud from "../hooks/useCrud";
 
 export type Punchline = {
   text: string;
