+++ frontend/src/pages/LobbyPage/index.tsx
@@ -8,7 +8,7 @@ import TextField from "../../components/TextField";
 import Dropdown from "../../components/Dropdown";
 import { Settings } from "../../GameRouter";
 import socket from "../../socket";
-import { PlayersContext } from "../../providers/ContextProviders/PlayersContextProvider";
+import { PlayersContext } from "../../contexts/players";
 
 const usePlayerIdState = createPersistedState("playerId");
